Reveal and hide the sideshow as workspaces are sliding out and back in.
Clip the sideshow so that portions of it don't show up behind workspaces as we enter and leave the expanded display mode.

Remove this._workspacesBackground from overlay. It is no longer useful because overlay background is no longer single-color.

Improve width assignment for the sideshow by keeping two variables: this._width which is the full width available for the sideshow in the regular mode and this._displayWidth which is the full width minus the padding we use on the left of the sideshow.